requirements.txt
bson==0.5.8
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
Django==2.2.14
django-cors-headers==3.4.0
djangorestframework==3.11.0
djongo==1.3.1
idna==2.10
numpy==1.19.1
pandas==1.0.5
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.10
